+++ libraries/resource_dnsmasq_local_app_rhel_yum.rb
@@ -2,7 +2,7 @@
 # frozen_string_literal: true
 
 # Cookbook Name:: dnsmasq-local
-# Library:: resource_dnsmasq_local_app_rhel
+# Library:: resource_dnsmasq_local_app_rhel_yum
 #
 # Copyright 2016, Socrata, Inc.
 #
@@ -19,14 +19,25 @@
 # limitations under the License.
 #
 
-require_relative 'resource_dnsmasq_local_app'
+require_relative 'resource_dnsmasq_local_app_rhel'
 
 class Chef
   class Resource
-    # A Dnsmasq package resouce specific to RHEL platforms.
+    # A Dnsmasq package resouce specific to RHEL7+, where we don't need
+    # the same hackery at uninstall time as with RHEL6.
     #
     # @author Jonathan Hartman <jonathan.hartman@socrata.com>
-    class DnsmasqLocalAppRhel < DnsmasqLocalApp
+    class DnsmasqLocalAppRhelYum < DnsmasqLocalAppRhel
+      provides :dnsmasq_local_app,
+               platform_family: 'rhel',
+               platform_version: '>= 7'
+
+      #
+      # Remove the Dnsmasq package.
+      #
+      action :remove do
+        package('dnsmasq') { action :remove }
+      end
     end
   end
 end