+++ spec/resources/dnsmasq_local_app/centos/7_3_1611_spec.rb
@@ -9,4 +9,13 @@ describe 'resources::dnsmasq_local_app::centos::7_3_1611' do
   let(:platform_version) { '7.3.1611' }
 
   it_behaves_like 'any CentOS platform'
+
+  context 'the :remove action' do
+    include_context description
+
+    it 'removes the dnsmasq package via yum' do
+      expect(chef_run).to remove_package('dnsmasq')
+        .with(options: nil)
+    end
+  end
 end