+++ spec/resources/dnsmasq_local_app/redhat/6_8_spec.rb
@@ -9,4 +9,13 @@ describe 'resources::dnsmasq_local_app::redhat::6_8' do
   let(:platform_version) { '6.8' }
 
   it_behaves_like 'any Red Hat platform'
+
+  context 'the :remove action' do
+    include_context description
+
+    it 'removes the dnsmasq package via rpm' do
+      expect(chef_run).to remove_rpm_package('dnsmasq')
+        .with(options: %w[--noscripts])
+    end
+  end
 end