+++ spec/resources/dnsmasq_local_app/redhat/7_3_spec.rb
@@ -9,4 +9,13 @@ describe 'resources::dnsmasq_local_app::redhat::7_3' do
   let(:platform_version) { '7.3' }
 
   it_behaves_like 'any Red Hat platform'
+
+  context 'the :remove action' do
+    include_context description
+
+    it 'removes the dnsmasq package via yum' do
+      expect(chef_run).to remove_package('dnsmasq')
+        .with(options: nil)
+    end
+  end
 end