+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Unreleased
 - Convert all the HWRPs to Chef custom resources
 - Patch the custom resources for Chef 13 compatibility
 - Support multiple config resources with different filename properties
+- Work around an failure condition with the :remove action on RHEL6
 
 v1.1.0 (2016-08-11)
 -------------------